refactor: 코드 리뷰 요청 전 리팩터링
