+++ src/js/components/Navbar/NavbarMain/Shipping.js
@@ -1,5 +1,5 @@
 import { Main } from '../../../Main.js';
-import { ShippingModal } from '../../element/Modal.js';
+import { ShippingModal } from '../../common/Modal.js';
 import { Component } from '/src/js/components/base/Component.js';
 
 export class Shipping extends Component {
